Added test for contact form validation. Moved contact form validation function as a utility function. Ran tests on validationForm.js
