--- __tests__/basicTest.test.js
@@ -1,5 +1,0 @@
-describe("Addition", () => {
-  it("knows that 2 and 2 make 4", () => {
-    expect(2 + 2).toBe(5);
-  });
-});
